Make ElementaryPredication not a subtype of Node.
This small change necessitated many changes throughout. Now there are no
longer cv or carg members of ElementaryPredication, and instead there are
properties that retrieve these from the argument list. This made some
serialization/deserialization tasks simpler.

The code is not fully working. In particular, the DMRX serialization is
producing bad links. I'm not sure if DMRX reading is working or not.