Make Mrs and Dmrs into functions, and fix unit tests.
There are many changes here. The initial change was to make Mrs() and
Dmrs() functions instead of subclasses of Xmrs, which necessitated some
changes, but I took the opportunity (perhaps to the detriment of an
atomic commit) to make Xmrs conform more to the API published on the
GitHub wiki (https://github.com/goodmami/pydelphin/wiki/delphin.mrs).