MRS codecs can now load/dump single MRSs.
The load(), loads(), dump(), and dumps() methods of the MRS codecs can now
take a 'single' keyword argument, as in:
  simplemrs.load('the-dog-barks.mrs', single=True)
This keyword argument will return the first parsed Xmrs instead of a
generator of Xmrses. For dump()/dumps(), the argument is expected to be a
single Xmrs instead of an iterator of them. For convenience, four
corresponding methods have been created that behave the same, but set
single to True. These methods are: load_one(), loads_one(), dump_one(),
and dumps_one().