SimpleMRS now allows surface strings.
This is a non-canonical extension to the format. The surface string
can also have a Lnk value, and they both appear before LTOP. So
semantically these are all the same, but the first two will get
their "surface" value assigned the value of the string.

[ "A dog barked."<0:17> LTOP: h0 INDEX: e2... ]
[ "A dog barked." LTOP: h0 INDEX: e2... ]
[ LTOP: h0 INDEX: e2... ]

Inner escaped quotes are fine, so you can have:

[ "Pat said, \"hi.\"" LTOP: h0 INDEX: e2... ]