doe.py now uses [incr tsdb()] profiles.
This change allows [incr tsdb()] profiles to be used as input
to doe.py with the `-i` option. The `--select`, `--apply`, and
`--filter` options can customize how the profiles are read. These
changes resulted in improvements to the `itsdb.TsdbProfile` class
such that it can take multiple filters or applicators that are
used when reading a table (with the read_table() method).