Fixed up the cascading filtering mechanism.
There was a variable scoping bug in the construction of key filters in delphin/itsdb.py, which is now fixed. This results in a more robust cascading filter. There is not the --cascade-filters option, which, if true, filters all dependent tables. For this commit, I also removed the Stag-like commands, as they are currently broken. They will be re-added when they are ready.