Output profiles not required; applicators bugfix.
Output profiles are not necessary for operations like --select, but the code was having validation errors when an output profile was not specified. This is no longer the case.

Also, the __init__ function of TsdbProfile had an incorrect identifier for the applicators, which caused an error. This has been fixed.