Add `join()` function to TsdbProfile and pyDelphin
This implements a naive join (nested loops) in order to join two tables in a profile based on a shared key (if it exists). The pyDelphin command then interfaces this through a --join option, which is only useful in conjunction with a --select operation. This resolves issue #21, though it would be nice to have a more efficient join operation, or one that can operate on non-key columns.