Significant performance improvement for itsdb join
The TsdbProfile.join() function naively performed a nested loop to look for joinable rows. Since most profiles are not huge (~25Mb for 1500 rows), and join only reads two tables, it makes sense to store one in memory. This allows for join to only read each table once, rather than N times, where N is the number of rows in table1. For a test profile (Tanaka Corpus profile tc-006), the time for a join and select went from 150 seconds to 1 second.

A small change to pyDelphin also helps print incremental results, rather than all at once.

And the sort_eps() is moved from delphin.mrs.ep to delphin.mrs, so it's available for pyDelphin. I'm not settled on this location, though.