Don't store EP objects; recreate them.
EPs and Nodes are now created just from info in the graph, which means that more info is stored in the graph itself. This rebuilding of objects can slow things down, but currently it helps when relabeling subgraphs.

Speaking of which, the relabel_nodes() methods on Xmrs and XmrsDigraph are added. On Xmrs, it is an in-place operation, but in XmrsDigraph, it returns a new graph.