Add commands to pyDelphin, and add mrs.compare
pyDelphin now takes commands, such as 'select', 'mkprof', and 'compare'. The 'select' command replaces the '--select' option, and options only relevant to 'select' are moved to the subparser. Similarly, the 'mkprof' command replaces the '--output' option, and options only relevant to 'mkprof' are moved to the subparser. The 'compare' command checks two itsdb profiles to see of the bags of MRSs for a parse-id are the same.

itsdb.py now has the match_rows() function, which returns groups of rows from two tables where a key is shared.

mrs.compare has a networkx-based isomorphism checker, and a compare_bags() function to look at two bags of MRSs, returning a triple of (unique-left, shared, unique-right).

The delphin/util.py module is added with a safe_int function, which casts to an int if it can, or else returns the original value. This is useful for numeric sorting.