Move XmrsDiGraph to xmrs.py and other small changes.
XmrsDiGraph is moved from mrs.util to mrs.xmrs and imports are adjusted accordingly.

The QEQ, LHEQ, and OUTSCOPES literals are put on HandleConstraint as class attributes, and the serializers and tests that use them have been changed accordingly.

The anchor parameter for ElementaryPredication is being phased out, so a new, optional, nodeid parameter supercedes it for now.