Sweeping changes to mrs.path. Fix Pred comparison.
mrs.path has significant changes to allow for rich paths (with context and nodeids). The
unit tests for paths are currently broken. This change also mostly gets rid of XmrsPath
objects, leaving only XmrsPathNode. Any functionality on XmrsPath that should be kept
should be moved to separate functions, I think.

Preds would raise an exception when being compared to None. The current fix is perhaps not general enough, but works for now.