Improve mrs.path performance and add BALANCED flag.
The BALANCED flag prevents a path node from using subpaths
individually instead of all-or-nothing. It only considers
local balance to subpaths (not the balance of the whole tree).
For example, given a path like:

    A(:ARG1/NEQ>B & :ARG2/NEQ>C)

The BALANCED flag will prevent these variations:

    A(:ARG1/NEQ> & :ARG2/NEQ>C)
    A(:ARG1/NEQ>B & :ARG2/NEQ>)

But these are still allowed:

    A(:ARG1/NEQ> & :ARG2/NEQ>)

Note that this flag is only relevant for the explore() method
(so perhaps it should be a function parameter instead of a
flag?)