Removing NetworkX core of Xmrs. Not complete.
This changes the NetworkX-based core of Xmrs to a custom datastructure, and moreover the component classes are now subclasses of namedtuples. This way, internal things (and serializers) use the tuple-based interface to get things, but the user can request the full objects and query things mostly as before.

This commit also changes Xmrs to use fewer properties and more methods to help indicate that the return values are copies of the internal structures.

Not everything is converted to the newer internals. Some things (like mrs.path) are probably broken, while others (like dmrx) are still relying on the "classy" interfaces, which are likely a bit slower.