Fix #46; Fix #48; Improve quantifier detection.
This commit changes Pred in the following ways:
  - the _relation suffix is no longer allowed as an alternate to _rel
  - if a predicate doesn't end in _rel, it is added during parsing, but not stored

These changes to Pred make it able to detect the POS for short forms like _the_q.
I also changed ElementaryPredication so that is_quantifier() looks for RSTR in its arguments instead of looking at the predicate. Node and Pred still look for a 'q' POS because I can't do much there.
There are probably a few places in xmrs.py that still need to be updated WRT quantifier detection.