Fix #49. Add TdlConsList and TdlDiffList classes.
The < > and <! !> syntax instantiates the TdlConsList and TdlDiffList classes, respectively, which are TdlDefinition subtypes with the addition of the `values()` method. This method returns list items as a Python list.

Note that this does not instantiate the list types when the lists are specified using the FIRST and REST (or LIST and LAST) features. Also, sublists are not lists (e.g. you cannot do `tdldef['REST'].values()`).

This commit also changes the behavior of empty cons lists (`< >`). Before, a `None` value was returned (so list items couldn't later be added), but now the behavior is the same as with an empty unbound list (`< ... >`). This is a side-effect of returning the TdlConsList class, and not necessarily an intentional change. It also occurred to me that there's nothing preventing *other* features from being defined on a list, such as `tdldef['REST.OTHER'] = "what's this?"`.