Validate XMRSs during deserialization; Fixes #53
SimpleMRS deserialization now takes an `errors=(ignore|warn|strict)` parameter. When the value is `strict` or `warn`, it runs Xmrs.validate() on the deserialized MRS. With `warn`, any resulting exception is converted to a warning; otherwise with `strict` the exception is re-raised.

This incurs a performance penalty on deserialization (~35%), but it can be avoided by setting `errors='ignore'`.

Currently this is only implemented for SimpleMRS. It may be added to other deserializers in the future.