Improve support for labeled trees.
This updates the Pegre parsing framework and includes a new S-Expression parser in delphin.util, which is used to parse labeled syntax trees (and, later, lisp-style output from ACE with --tsdb-stdout). The EDS module is also updated to use some of the new Pegre features.

The delphin.interfaces.base module now includes a tree() method to either parse a stored labeled tree, or extract one from a JSON-serialized derivation containing the labels. The derivation module is correspondingly modified to be able to produce these embedded labels when the to_dict() method is used.