Require setuptools<36.0.0 for testing; fix tests
Setuptools 36.0.0 has an undeclared dependency to six; pin it to <36.0.0.
Simply not upgrading pip (the last commit) didn't work because it also didn't update pytest, which is necessary.

Also, in fixing these, I uncovered some broken tests which I addressed, and fixed some things (but not all) that caused warnings.