Add Relations class and update Field class
Add itsdb.Relations, which allows for reading/writing a relations file
to strings as well as (reading from) files.

Change the 'other' attribute of itsdb.Field to 'partial', which is just
a simple boolean value, since that's all 'other' is used for.

Also, add a few more patterns to .gitignore so editors stop indexing
.tox/, etc.

Resolves #99