Remove pandas dependency for itsdb, add features
Pandas was a good fit for handling the tabular [incr tsdb()] data, but
unfortunately it is not supported for Python 3.3 or 3.4, whereas
PyDelphin is, so I chose to not use it at this time.

This commit implements most of the required functionality for the
replacement classes and functions and adds unit tests, but there is
still some functionality missing, as well as docstrings and other
documentation.

Resolves #58
Resolves #89
Resolves #101
Resolves #102