Output conformant DMRX and MRX when POS is missing
The previous fix allowed PyDelphin to deal with missing POS values in
Preds, but it still output DMRX that had no pos attribute on nodes,
which was non-conformant with the DTD. This commit makes pos=""
equivalent to a None value of pos for DMRX and MRX. MRX is still
non-conformant in that it won't output <label> or <var> as TOP and
INDEX when they are None on the Xmrs object, but there is not a good
solution for that case. This commit also allows MRX to deal with
missing TOP and INDEX on input.

Addresses #129