Capture and record 'run' information with ACE
Each time an ACE process is opened, various system information is
collected so that it can be stored later (e.g., in a [incr tsdb()]
profile), including the operating system version, host and user names,
and the process start time. Some information is collected after the
process ends, so this information should not be written to a profile
until the process is complete.