Validate ACE arguments against whitelist
This creates an `ArgumentParser` that mocks a subset of ACE's
command-line interface in order to create a whitelist of options that
are not likely to cause problems with PyDelphin's ACE interface. As a
bonus, it also validates argument types (e.g., that the value of the
`-n` option is an integer). Because this prevents some options from
being sent to ACE that could have been used prior, this is a breaking
change, but those options would have likely interfered with
PyDelphin's parsing of ACE's output, so it is unlikely that this
change will break any *working* code.

Resolves #149