Add 'ace' codec for the 'convert' subcommand
This change makes it easier to convert the MRS output of ACE without
having to massage the data stream first. It works with normal ACE
output and --tsdb-stdout output. For the former, it uses the SENT:
line, if available, to set the 'surface' attribute of the following
MRSs (until the next double-newline, indicating the end of the results
list).

Resolves #92