Add new-style TDL parsing.
This adds a lot of code to tdl.py, although much of the old stuff will
be removed in a future release. The new-style parsing is ~36% slower
at reading the ERG's lexicon, but it is better able to deal with
malformed TDL, and it handles docstrings and comments in all valid
places.

Addresses #153, #167, #168, and #170