Reduce Conjunction creation; add docs; fix tests
This changes the API a bit to allow for more bare-Terms as AVM values,
which makes traversing the structure easier in a sense (though perhaps
less predictable/automatable).

Docstrings are added to all public classes and functions, and the reST
documentation files were updated.

Tests were updated to the above changes, as well as an attempt to use
tox on the CI server.

At this point the functionality resolves several issues:

Resolves #82
Fixes #153
Resolves #168
Fixes #170