Fix itsdb.Table uncommitted record retrieval
Before only records for the number of lines in the file would be
returned, so any additional records would not be retrieved. I think
the opposite problem is still there: when the in-memory table is
smaller than the file.

This commit also allows itsdb.make_row() to take tuples to be made
into rows. Before only things with `get()` methods would work (like
dicts or Records).