Re-enable append param for itsdb.TestSuite.write()
Looks like it was lost along the way. It is re-added but needs
testing.

This commit also ensures that the computation of gzipped/regular table
filenames works with strings or with Path-like objects, and fixes a
check when writing tables to make sure we aren't treating a file like
a directory.