Implement streaming TestSuite processing
This modifies how tables are written and fixes some nasty bugs in
order to get streaming (incremental) processing of TestSuites. The
worst bug was when attempting to overwrite a table using the table's
rows, because the previous work on incremental reading meant that the
file would be overwritten as it was being read. Now temporary files
are written to first, then copied (or appended) to the destination.

Currently this only works for Python 3. Python 2 has good ol'
UnicodeDecodeError when trying to make strings out of ACE output. I'm
not certain that this wasn't already a problem, though.

Addresses #186