Change misplaced property error to warning
And make SEM-I parsing a bit more robust to, e.g., a dot in a
predicate name.

Resolves #199

But this does not relax tests on compatible property values, etc., so
in order to load the ERG's SEM-I (as of 2019-04-11) you need to change
MOOD's value to 'mood' and rename the 'person' property to, e.g.,
'pers' so it doesn't collide with the predicate 'person'. But note the
following ERG issue for the latter point:

    https://github.com/delph-in/erg/issues/14

So we should maybe have separate hierarchies for properties,
variables, and predicates.