Begin implementing new MRS and SimpleMRS
This breaks unit tests because there are too many changes to
reasonably fit into a single commit, but this is a good starting
point. It starts creating the new classes in delphin.mrs._mrs and the
simplemrs module now uses that, but the output is no longer an Xmrs so
many things break.

This commit also changes delphin.variable.sort() to type() with sort()
as an alias. I call them variable types in many places so I'm trying
to be consistent.

I also added a Lnk.default() method which creates an empty Lnk. The
goal here is to not have None a valid value for lnk on *MRS objects,
then I don't have to check for None, e.g., on serialization.