Big structural change to semantics hierarchy
Still working out the class hierarchy.

This commit also removes delphin.eds.Edge, since edges in EDS are
often grouped with the nodes.

This also includes a bit of half-baked implementations of semantic
operations in delphin.semops.

This commit also adds some forgotten files:

 - delphin.hierarchy (and docs and tests)
 - delphin.sembase docs

(note, CHANGELOG.md was not updated for this commit; need to do it
later)