Refine scope module; get representatives() working
I think neither range() nor domain() are appropriate names for the
scope function as I had it, as these terms mean something specific in
the literature regarding quantifiers and the function did something
different. In the end I re-imagined the function as returning the
partial scope trees that you get for MRSs, so the function is now
called tree_fragments(). In addition, the new UnderspecifiedScope
class makes it pretty straightforward to interact with these, such as
determining if a predication exists within a scope. These changes make
the calculation of scope representatives almost trivial.

I also abstracted the sorting of representatives to a priority
function (similar to delphin.sembase.role_priority()), although the
function depends on the data in an MRS, so the new function
make_representative_priority() returns a closure over the input MRS.
Having sorting be a separate function makes it possible to insert more
EDS-like sorting in the future.