Make delphin a namespace pkg, give others versions
As long as each module or package in delphin/ is included by default,
it should use PyDelphin's version taken from delphin.__about__. If a
module or package is spun off to be separate, it should get a new
version starting from the last PyDelphin version it used.

This commit also removes the 'extra' package from the install
metadata (but not the package itself).