Simplify itsdb API
This removes a lot of complexity in the TestSuite (and related
classes) API and makes it more like a DBAPI2 implementation.
There is no more "writing" of Tables or TestSuites, only "committing"
at the TestSuite level. TestSuites are always backed by files, even if
they are temporary files that get deleted when the TestSuite is
garbage collected.

Resolves #228
Resolves #230
Resolves #233

The delphin.tsql and delphin.commands modules still need to be updated
to the new APIs.