Ensure TSDB dates are TSDB-formatted on output
If a date is not parsed (e.g., when reading with tsdb.Database), no
changes are made, but if the value is parsed (e.g., with
itsdb.TestSuite) then the date will be written in a TSDB-compatible
format, namely DD-MM-YYYY with English month abbreviations ('jan',
'feb', etc.) to avoid ambiguity. Since it is impossible to have a
Python datetime.datetime object without a day-of-the-month, it is not
possible to output the shortened MM-YYYY format, and furthermore years
are always output with 4 digits. Times are only printed if they are
not 00:00:00 (which is the default when parsing if no time was
specified). Before, dates were output according to the YYYY-MM-DD ISO
date format, which LOGON's `tsdb` command would not parse.

These formatting rules are made available via the new `tsdb.format()`
function.

This commit also changes tsdb.Table.select() to use
Table.column_index() instead of directly querying the internal data
structures in case a subclass wants to redefine Table.column_index().