Rename tsdb components to follow TSDB terminology.
Renaming things helps to avoid name confusion with itsdb components.
This means that tables are now relations (hopefully not confusable
with the previously misnamed Relation class which was part of the
schema description), and rows are records.

Some things are also renamed for conciseness.  Things like
open_table() and write_table() are now just open() and write(), but
write_database() remains as is. Also, encode_row() and decode_row()
are just encode() an decode(), but make_row() becomes make_record().

This commit also updates the documentation and tests to correspond to
these changes.