Make tsdb.write_database() work with missing files
This tests the behavior that `tsdb.Database(...)[name]` raises a
TSDBError not only if `name` is not defined by the schema but also if
it is defined but the file is missing. It also changes
tsdb.write_database() to detect this case and write an empty file (if
it should write `name` at all).