Disallow appending to gzipped files in tsdb.write
Even if a user requested gzip=True, there were several conditions to
determine if the output should be gzipped, such as if the resultgin
file were empty, which is hard to determine. If gzip=None, it
dependended also on the presence of an existing gzipped file. In the
end, I thought the complicated conditions were not very maintainable
and also the behavior was a bit "magic", so I opted instead for a
simpler approach. Now, the gzip parameter defaults to False instead of
None, and gzip=True may not be used with append=True or if appending
to an existing gzipped file (that is, even if gzip=False). If writing
a new file, gzip=True will compress if the file is non-empty.