Add tests for mkprof --refresh; fix resulting bugs
The change is quite large because I also streamlined some use cases,
changed some errors from mutually exclusive parameters to just
ignoring one (and this is documented), and making sure things get
cleaned up afterwards.

A bug with refreshing was that it was not remaking the records with
the destination database's schema.