Fix intermittent test failure with TSQL
If a series of joins happened in a weird order the appropriate field
index was not getting added. E.g., 'select i-id mrs' selects the
qualified names 'item.i-id' and 'result.mrs', but if the joins were
result+parse+item instead of item+parse+result, then the qualified
name parse.i-id would be in the index instead of item.i-id. Now both
are added, pointing to the same thing.