Update delphin.scope for predicate-level scoping
This removes scope.tree_fragments() and scope.UnderspecifiedScope in
favor of updates to the ScopeMap type and adding
scope.descendants(). The reason is so scope.representatives() can get
predicate-granularity scope descendants, which was useful for
excluding representative candidates. The other stuff is removed
because it is no longer used, although it may still be useful.

See #241

This update also tweaks the representative sorting algorithm to
include TENSE information. Diffs on the EDS task with the LKB are now
minimal.