Fix delphin.commands.process with new *tsdb code
This apparently wasn't tested as broke. Now it is fixed, but as
TestSuite.process() no longer accepts Tables as a 'source' parameter,
there was no direct way to filter inputs with TSQL. Now a temporary
directory is created with the filtered results, and it is cleaned up
after processing is complete.