Join TSQL selections as raw, compare as cast
When joining and filtering tables in TSQL queries, only the join keys
and rows used in conditions are cast to their datatypes, and all data
stored in the resulting Selection object is raw. This also has the
benefit that a normal tsql.select() will return tuples of raw strings,
but passing record_class=itsdb.Row will let them be auto-cast.