Handle incomplete output from ACE w/ --tsdb-stdout
This captures the situation where the parser tries to read past the
end of the string. It's not guaranteed to work in the case that ACE is
killed after printing a partial but valid set of s-expressions, in
which case it will just record the partial information and move on
without any other warning. I'm not sure yet how to detect this latter
situation.

Fixes #237