Use absolute URLs in README; remove gTest mention
Relative links in the README do not work when the README is published
elsewhere (e.g., PyPI), so now use absolute links.

And gTest will probably not be updated for PyDelphin v1.0.0, so its
mention in the README is removed.

This commit also makes a note in CHANGELOG.md so it's more clear that
the latest release is not grouped under "Unreleased".