Make commands.convert tests do reverse direction
For some reason the bidirectional conversion testing was only
converting src to tgt and tgt to tgt, apparently to avoid EDS to
non-EDS conversion. This made more sense when MRS and DMRS shared the
Xmrs representation, but since that's gone it was hiding bugs. Now it
does bidirectional conversions properly, and EDS conversions are not
attempted as bidirectional.