Use the regex library with REPP if available
This commit also adds the `delphin.repp.REPPWarning` class which warns
if (a) the regex library is not used, and (b) if the library is used
and it fails to compile a pattern potentially containing unescaped
brackets in a character class.

This adds the `repp` extra for pip installations, and makes it part of
`tests`. If the regex library is not installed, one test is skipped.