Re-ensure the public API is fully documented
Some exceptions did not have autodoc stubs, but now they
do. Documentation coverage shows a few classes or methods that are
still undocumented:

- delphin.tsdb.Relation
- delphin.tsql.Selection
- delphin.web.server.ProcessServer (.on_get(), .spawn())
- delphin.web.server.TestSuiteServer (.on_get(), .on_get_name(),
  .on_get_table())

... but I don't think these should be documented. `Relation` is mainly
used as a type currently, and `Selection` should possibly use a
non-public name. The web server class methods are not documented
because doing so might over-commit to the specific (currently
Falcon-based) implementation, and the users are not expected to call
those methods.