Remove networkx dependency
The delphin.util._isomorphism() function is added and it is used by
delphin.mrs.is_isomorphic(), which removes the need for the NetworkX
dependency. NetworkX was the largest dependency---many times bigger
than PyDelphin itself---and it was only used for a single
function. The new solution also implements the same algorithm (VF2),
but I implemented it without consulting the NetworkX code, and it is
much less flexible but is fine for PyDelphin's purposes. It passed the
old tests (and some new ones) and is just as fast.