Simplify XML codec tests
Python 3.8 preserves attribute order because of ordered dictionaries,
but previous versions sorted attributes alphabetically (I think), so
it is no longer feasible to test the string output of XML
serialization. Instead I simplified the DMRX and MRX tests to just
round-trip a DMRS/MRS object and compare that it's the same.