Add final newlines to test profile data
Writing with append breaks if the file doesn't end with a newline. One
fix (untested) is to check that the last character of the copied
portion of an append is a newline and insert one if not. With this
change, however, I am just saying that files not ending in newlines
are not fully supported.