Add 'source' parameter to itsdb.FieldMapper
This parameter is not needed so much in parsing, but for generation
there was an issue where the input relation only has the 'parse-id'
key so the original 'i-id' was not associated with the results. Then,
queries like `delphin select 'i-id surface'` would fail. Now the
source parameter, if given, can reestablish the association.

Fixes #283