Fix column names with delphin mkprof --delimiter
The final newline wasn't being stripped off the header line, so when
it tried to get the column names, the last one had a `\n` at the end,
which meant it did not match up to the actual column name (e.g.,
`i-author\n` instead of `i-author`). Now the newline is stripped.

Fixes #288