Fix backtracking in isomorphism check
An overly defensive condition on a while-loop caused the algorithm to
abort when an intermediate state had no further candidates instead of
backtracking again to the previous state. Now the loop condition only
checks if the mapping is complete, but other conditions within the
loop break out when there are no further alternatives.

In addition, the iteration of the set of candidates was not as
consistent as I thought it would be, leading to flaky tests. The
candidates are now sorted. Furthermore, they are sorted in reverse
order so in the (common) case where the two MRSs are exactly the same
the first candidate will be the correct one.