Improve the error message on TDL RecursionError
I expect the situation to be so rare that I'd rather just punt and let
the user handle it. The solution is simple, so I make PyDelphin give
an error message to explain what to do.

Of course, this doesn't *really* solve the problem if they have a TDL
list with an obscene number of items that goes beyond the stack limit,
but I'll wait for the problem to arise before I try to solve it.

Resolves #294